replace mono/ana with Crystal
