+++ src/tavi/library/Instrument/instrument.py
@@ -7,10 +7,9 @@ Initially configured to handle HFIR instruments.
 
 from typing import Literal, Optional
 
-from tavi.library.component.analyzer import Analyzer
 from tavi.library.component.collimators import Collimators
+from tavi.library.component.crystal import Crystal
 from tavi.library.component.goniometer import Goniometer
-from tavi.library.component.monochromater import Monochromator
 
 
 class Instrument:
@@ -18,8 +17,8 @@ class Instrument:
 
     def __init__(
         self,
-        monochromator: Optional[Monochromator] = None,
-        analyzer: Optional[Analyzer] = None,
+        monochromator: Optional[Crystal] = None,
+        analyzer: Optional[Crystal] = None,
         collimators: Optional[Collimators] = None,
         goniometer: Optional[Goniometer] = None,
         mode: Literal["fix_ei", "fix_ef"] = "fix_ef",
@@ -47,22 +46,3 @@ class Instrument:
         else:
             self.ef = self.ei - e
         return self.ei, self.ef
-
-    # def get_two_theta(self, hkl: tuple, sample: Sample, ei: float, ef: float) -> float:
-    #     """Get two_theta."""
-    #     ki = SE2K(ei)
-    #     kf = SE2K(ef)
-    #     q_norm = sample.ol.q_norm_from_hkl(hkl)
-    #     two_theta_rad = get_angle_from_triangle(ki, kf, q_norm)
-    #     sign = 1 if self.goni.s2_sense == "+" else -1
-    #     return two_theta_rad * sign
-
-    # def get_psi(self, hkl: tuple, sample: Sample, ei: float, ef: float) -> float:
-    #     """Get psi. Angle between ki and Q."""
-    #     ki = SE2K(ei)
-    #     kf = SE2K(ef)
-    #     q_norm = sample.ol.q_norm_from_hkl(hkl)
-    #     psi_rad = get_angle_from_triangle(ki, q_norm, kf)
-    #     # sign of psi is always opposite of s2
-    #     sign = -1 if self.goni.s2_sense == "+" else 1
-    #     return psi_rad * sign
