--- src/tavi/library/component/analyzer.py
@@ -1,62 +1,0 @@
-"""Analyzer."""
-
-from typing import Literal
-
-import numpy as np
-
-from tavi.library.component.crystal import crystal_d
-from tavi.library.experiment.utilities import SE2K
-
-
-class Analyzer:
-    """Analyzer that holds horizontal and vertical mosaic."""
-
-    def __init__(
-        self, mosaic_h: float = 30.0, mosaic_v: float = 30.0, crystal: str = "PG002", sense: Literal["+", "-"] = "-"
-    ) -> None:
-        """Init."""
-        self._mosaic_h = mosaic_h
-        self._mosaic_v = mosaic_v
-        self.crystal = crystal
-        self.d_spacing = crystal_d.get(crystal, 0)
-        self.sense = sense
-
-    @property
-    def mosaic_h(self) -> float:
-        """Getter."""
-        return np.radians(self._mosaic_h / 60)
-
-    @mosaic_h.setter
-    def mosaic_h(self, val: float) -> None:
-        """Setter."""
-        self._mosaic_h = val
-
-    @property
-    def mosaic_v(self) -> float:
-        """Getter."""
-        return np.radians(self._mosaic_v / 60)
-
-    @mosaic_v.setter
-    def mosaic_v(self, val: float) -> None:
-        """Setter."""
-        self._mosaic_v = val
-
-    @property
-    def sense(self) -> int:
-        """Get mono sense."""
-        return 1 if self._sense == "+" else -1
-
-    @sense.setter
-    def sense(self, val: str) -> None:
-        """Set sense."""
-        self._sense = val
-
-    def theta_a(self, ei: float) -> float:
-        """
-        Calculate scattering angle based on Bragg's law.
-
-        2dsin(theta) = lambda = 2 pi / k, k is wavevector.
-        """
-        kf = SE2K(ei)
-        asin = np.pi / (self.d_spacing * kf)
-        return np.arcsin(asin)
